crm/crm_app/serializers/global_var_serializers.py
from rest_framework import serializers
from ..models import GlobalVarModel


class GlobalVarSerializer(serializers.ModelSerializer):
    class Meta:
        model = GlobalVarModel
        fields = '__all__'