crm/requirements.txt
amqp==5.1.0
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
joblib==1.1.0
kombu==5.2.4
numpy==1.22.3
packaging==21.3
prompt-toolkit==3.0.28
psycopg2==2.9.3
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tmdbv3api==1.7.6
tzdata==2021.5
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wordsegment==1.3.1
wrapt==1.14.0
